+++ Changes.md
@@ -5,6 +5,7 @@ API
 ---
 
 - ThreadState : Added `process()` method.
+- Process : Added const overload for `handleException()` method. The non-const version will be removed in future.
 
 1.3.2.0 (relative to 1.3.1.0)
 =======